requirements.txt
# Mutual Funds RAG — all phases (run locally)
# 1) Copy .env.example to .env and paste your GROQ_API_KEY
# 2) pip install -r requirements.txt
# 3) python run_local.py --refresh   (first time: scrape + process + index)
# 4) python run_local.py             (start chat UI)

python-dotenv>=1.0.0

# Phase 1
beautifulsoup4>=4.12.0
playwright>=1.40.0
requests>=2.31.0

# Phase 2 (stdlib only; optional: pandas, pyarrow for Parquet)

# Phase 3
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Phase 4 (uses Phase 3)

# Phase 5
groq>=0.4.0

# Phase 6
streamlit>=1.28.0

# Phase 7 (uses Phases 1–3, 6)
# no extra deps

# Dev/tests
pytest>=7.0.0
